[develop] 我的收藏页-收藏网址列表功能、收藏链接Web页面

+++ lib/model/collect_link_model.dart
@@ -2,20 +2,20 @@ import 'package:flutter_wan_android_getx/generated/json/base/json_convert_conten
 import 'package:get/get.dart';
 
 class CollectLinkModel with JsonConvert<CollectLinkModel> {
-  String? desc;
-  String? icon;
-  int? id;
-  String? link;
-  String? name;
-  double? order;
-  int? userId;
-  double? visible;
-  bool? collect;
+	String? desc;
+	String? icon;
+	double? id;
+	String? link;
+	String? name;
+	double? order;
+	double? userId;
+	double? visible;
+	bool? collect;
 
-  /// 可观察变量 isCollect 是否收藏
-  final Rx<bool> _isCollect = false.obs;
+	/// 可观察变量 isCollect 是否收藏
+	final Rx<bool>_isCollect = false.obs;
 
-  get isCollect => _isCollect.value;
+	get isCollect => _isCollect.value;
 
-  set isCollect(value) => _isCollect.value = value;
+	set isCollect(value) => _isCollect.value = value;
 }
